# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface UniversityCredentials {
  /** @minLength 1 */
  username: string;
  /** @minLength 1 */
  password: string;
}

export interface UniversityLoginResult {
  success: boolean;
  /** @nullable */
  sessionToken?: string | null;
  /** @nullable */
  sesskey?: string | null;
  /** @nullable */
  userFullname?: string | null;
  /** @nullable */
  error?: string | null;
}

export interface UniversityTask {
  id: string;
  title: string;
  courseName: string;
  /** @nullable */
  dueDate?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  url?: string | null;
}

export interface UniversityTaskList {
  tasks: UniversityTask[];
}

export interface UniversityError {
  error: string;
}

export interface AcademicStudent {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  career?: string | null;
  /** @nullable */
  group?: string | null;
  /** @nullable */
  generation?: string | null;
  /** @nullable */
  overallAverage?: number | null;
}

export interface AcademicUnitGrade {
  number: number;
  /** @nullable */
  grade?: number | null;
  /** @nullable */
  gradeRaw?: string | null;
  /** @nullable */
  modality?: string | null;
  /** @nullable */
  weight?: number | null;
}

export interface AcademicSubject {
  id: string;
  name: string;
  /** @nullable */
  code?: string | null;
  /** @nullable */
  teacher?: string | null;
  /** @nullable */
  finalGrade?: number | null;
  /** @nullable */
  finalGradeRaw?: string | null;
  units: AcademicUnitGrade[];
}

export interface AcademicPeriod {
  id: string;
  label: string;
  /** @nullable */
  year?: string | null;
  /** @nullable */
  average?: number | null;
  subjects: AcademicSubject[];
}

export interface AcademicSnapshot {
  student: AcademicStudent;
  periods: AcademicPeriod[];
  lastUpdatedAt: string;
  source?: string;
}

export interface UniversityGradesResult {
  success: boolean;
  /** @nullable */
  error?: string | null;
  data?: AcademicSnapshot | null;
}

export type GetUniversityTasksParams = {
sessionToken: string;
sesskey: string;
};