# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: university
    description: University Moodle platform integration
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /university/login:
    post:
      operationId: universityLogin
      tags: [university]
      summary: Login to the university Moodle platform
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UniversityCredentials"
      responses:
        "200":
          description: Login result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UniversityLoginResult"

  /university/tasks:
    get:
      operationId: getUniversityTasks
      tags: [university]
      summary: Get upcoming assignments from the university Moodle platform
      parameters:
        - name: sessionToken
          in: query
          required: true
          schema:
            type: string
        - name: sesskey
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of upcoming university tasks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UniversityTaskList"
        "401":
          description: Session expired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UniversityError"

  /university/grades:
    post:
      operationId: getUniversityGrades
      tags: [university]
      summary: Read academic grades from the UTXJ SII
      description: Logs into the institutional SII with the same university credentials and returns normalized academic data.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UniversityCredentials"
      responses:
        "200":
          description: Academic data from SII
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UniversityGradesResult"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    UniversityCredentials:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 1

    UniversityLoginResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        sessionToken:
          type: ["string", "null"]
        sesskey:
          type: ["string", "null"]
        userFullname:
          type: ["string", "null"]
        error:
          type: ["string", "null"]

    UniversityTask:
      type: object
      required: [id, title, courseName]
      properties:
        id:
          type: string
        title:
          type: string
        courseName:
          type: string
        dueDate:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        url:
          type: ["string", "null"]

    UniversityTaskList:
      type: object
      required: [tasks]
      properties:
        tasks:
          type: array
          items:
            $ref: "#/components/schemas/UniversityTask"

    UniversityError:
      type: object
      required: [error]
      properties:
        error:
          type: string

    UniversityGradesResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        error:
          type: ["string", "null"]
        data:
          anyOf:
            - $ref: "#/components/schemas/AcademicSnapshot"
            - type: "null"

    AcademicSnapshot:
      type: object
      required: [student, periods, lastUpdatedAt]
      properties:
        student:
          $ref: "#/components/schemas/AcademicStudent"
        periods:
          type: array
          items:
            $ref: "#/components/schemas/AcademicPeriod"
        lastUpdatedAt:
          type: string
        source:
          type: string

    AcademicStudent:
      type: object
      properties:
        name:
          type: ["string", "null"]
        career:
          type: ["string", "null"]
        group:
          type: ["string", "null"]
        generation:
          type: ["string", "null"]
        overallAverage:
          type: ["number", "null"]

    AcademicPeriod:
      type: object
      required: [id, label, subjects]
      properties:
        id:
          type: string
        label:
          type: string
        year:
          type: ["string", "null"]
        average:
          type: ["number", "null"]
        subjects:
          type: array
          items:
            $ref: "#/components/schemas/AcademicSubject"

    AcademicSubject:
      type: object
      required: [id, name, units]
      properties:
        id:
          type: string
        name:
          type: string
        code:
          type: ["string", "null"]
        teacher:
          type: ["string", "null"]
        finalGrade:
          type: ["number", "null"]
        finalGradeRaw:
          type: ["string", "null"]
        units:
          type: array
          items:
            $ref: "#/components/schemas/AcademicUnitGrade"

    AcademicUnitGrade:
      type: object
      required: [number]
      properties:
        number:
          type: integer
        grade:
          type: ["number", "null"]
        gradeRaw:
          type: ["string", "null"]
        modality:
          type: ["string", "null"]
        weight:
          type: ["number", "null"]