# lib/api-zod/src/generated/types/academicPeriod.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AcademicSubject } from './academicSubject';

export interface AcademicPeriod {
  id: string;
  label: string;
  /** @nullable */
  year?: string | null;
  /** @nullable */
  average?: number | null;
  subjects: AcademicSubject[];
}