# lib/api-zod/src/generated/types/academicSnapshot.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AcademicPeriod } from './academicPeriod';
import type { AcademicStudent } from './academicStudent';

export interface AcademicSnapshot {
  student: AcademicStudent;
  periods: AcademicPeriod[];
  lastUpdatedAt: string;
  source?: string;
}