# lib/api-zod/src/generated/types/academicStudent.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AcademicStudent {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  career?: string | null;
  /** @nullable */
  group?: string | null;
  /** @nullable */
  generation?: string | null;
  /** @nullable */
  overallAverage?: number | null;
}