# lib/api-zod/src/generated/types/academicSubject.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AcademicUnitGrade } from './academicUnitGrade';

export interface AcademicSubject {
  id: string;
  name: string;
  /** @nullable */
  code?: string | null;
  /** @nullable */
  teacher?: string | null;
  /** @nullable */
  finalGrade?: number | null;
  /** @nullable */
  finalGradeRaw?: string | null;
  units: AcademicUnitGrade[];
}