# lib/api-zod/src/generated/types/academicUnitGrade.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AcademicUnitGrade {
  number: number;
  /** @nullable */
  grade?: number | null;
  /** @nullable */
  gradeRaw?: string | null;
  /** @nullable */
  modality?: string | null;
  /** @nullable */
  weight?: number | null;
}