# lib/api-zod/src/generated/types/universityGradesResult.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AcademicSnapshot } from './academicSnapshot';

export interface UniversityGradesResult {
  success: boolean;
  /** @nullable */
  error?: string | null;
  data?: AcademicSnapshot | null;
}